refactor: page.tsx -> page.client.tsx 변경 후 page.tsx에서 import
